+++ src/message/props.ts
@@ -2,7 +2,6 @@
 
 /**
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
- * updated at 2021-12-12 19:17:30
  * */
 
 import { TdMessageProps } from './type';
@@ -33,9 +32,12 @@ export default {
     type: String as PropType<TdMessageProps['theme']>,
     default: 'info' as TdMessageProps['theme'],
     validator(val: TdMessageProps['theme']): boolean {
+      if (!val) return true;
       return ['info', 'success', 'warning', 'error', 'question', 'loading'].includes(val);
     },
   },
+  /** 关闭消息时触发 */
+  onClose: Function as PropType<TdMessageProps['onClose']>,
   /** 当关闭按钮存在时，用户点击关闭按钮触发 */
   onCloseBtnClick: Function as PropType<TdMessageProps['onCloseBtnClick']>,
   /** 计时结束后触发 */